Rename basic.js to page.js
